Magik.Params (Magik v0.8.0)

Params provide some helpers method to work with parameters

Functions

clean_nil(param)

Specs

clean_nil(any()) :: any()

Clean all nil field from params, support nested map and list.

Example

params = %{"keyword" => nil, "email" => nil, "type" => "customer"}
Magik.Params.clean_nil(params)
# => %{"type" => "customer"}

params = %{user_ids: [1, 2, nil]}
Magik.Params.clean_nil(params)
# => %{user_ids: [1, 2]}

plug_srub(conn, keys \\ [])

A plug which do srubbing params

Use in Router

defmodule MyApp.Router do
  ...
  plug Magik.Params.plug_srub
  ...
end

Use in controller

plug Magik.Params.plug_srub when action in [:index, :show]
# or specify which field to scrub
plug Magik.Params.plug_srub, ["id", "keyword"] when action in [:index, :show]

scrub_param(struct)

Convert all parameter which value is empty string or string with all whitespace to nil. It works with nested map and list too.

Example

params = %{"keyword" => "   ", "email" => "", "type" => "customer"}
Magik.Params.scrub_param(params)
# => %{"keyword" => nil, "email" => nil, "type" => "customer"}

params = %{user_ids: [1, 2, "", "  "]}
Magik.Params.scrub_param(params)
# => %{user_ids: [1, 2, nil, nil]}