Magik.Validator (Magik v0.10.0)

Some helpers function to do validate data

Validate type

validate inclusion/exclusion

validate length for string and enumerable types

validate number

validate string format/pattern

validate custom function

validate allow_nil or not

Each of these validations can be used separatedly

iex(2)>   Magik.Validator.validate_type(10, :integer)
:ok
iex(3)>   Magik.Validator.validate_type(10, :string)
{:error, "is not a string"}
iex(3)>   Magik.Validator.validate_number(9, [min: 10, max: 20])
{:error, "must be greater than or equal to 10"}

Or you can combine multiple condition at one

iex(12)> Magik.Validator.validate(10, type: :integer, number: [min: 10, max: 20])
:ok
iex(13)> Magik.Validator.validate("email@g.c", type: :string, format: ~r/.+@.+.[a-z]{2,10}/)
{:error, "format not matched"}

Types

error()

Specs

error() :: {:error, String.t()}

support_length_types()

Specs

support_length_types() :: String.t() | map() | list() | tuple()

Functions

validate(value, validators)

Specs

validate(any(), keyword()) :: :ok | error()

Validate value against list of validations.

iex(13)> Magik.Validator.validate("email@g.c", type: :string, format: ~r/.+@.+.[a-z]{2,10}/)
{:error, "format not matched"}

All supported validations:

type: validate datatype

format: check if binary value matched given regex

number: validate number value

length: validate length of supported types. See validate_length/2 for more details.

in: validate inclusion

not_in: validate exclusion

func: custom validation function follows spec func(any()):: :ok | {:error, message::String.t()}

validate_embed(value, embed_type)

Validate embed types

validate_exclusion(value, enum)

Check if value is not included in the given enumerable. Similar to validate_inclusion/2

validate_format(value, check)

Specs

validate_format(String.t(), Regex.t()) :: :ok | error()

Checks whether a string match the given regex.

iex(11)> Magik.Validator.validate_format("year: 2001", ~r/year: {4}/)
:ok
iex(12)> Magik.Validator.validate_format("hello", ~r/+/)
{:error, "does not match format"}

validate_inclusion(value, enum)

Check if value is included in the given enumerable.

iex(21)> Magik.Validator.validate_inclusion(1, [1, 2])
:ok
iex(22)> Magik.Validator.validate_inclusion(1, {1, 2})
{:error, "given condition does not implement protocol Enumerable"}
iex(23)> Magik.Validator.validate_inclusion(1, %{a: 1, b: 2})
{:error, "not be in the inclusion list"}
iex(24)> Magik.Validator.validate_inclusion({:a, 1}, %{a: 1, b: 2})
:ok

validate_length(value, checks)

Specs

validate_length(support_length_types(), keyword()) :: :ok | error()

Check if length of value match given conditions. Length condions are the same with validate_number/2

iex(15)> Magik.Validator.validate_length([1], min: 2)
{:error, "length must be greater than or equal to 2"}
iex(16)> Magik.Validator.validate_length("hello", equal_to: 5)
:ok

Supported types

list

map

tuple

keyword

string

validate_number(value, checks)

Specs

validate_number(integer() | float(), keyword()) :: :ok | error()

Validate number value

iex(3)> Magik.Validator.validate_number(12, min: 10, max: 12)
:ok
iex(4)> Magik.Validator.validate_number(12, min: 15)
{:error, "must be greater than or equal to 15"}

Support conditions

equal_to

greater_than_or_equal_to | min

greater_than

less_than

less_than_or_equal_to | max

validate_number(x, [min: 10, max: 20])

validate_type(value, struct_name)

Validate data types.

iex(1)> Magik.Validator.validate_type("a string", :string)
:ok
iex(2)> Magik.Validator.validate_type("a string", :number)
{:error, "is not a number"}

Support built-in types:

boolean

integer

float

number (integer or float)

string | binary

tuple

map

array

atom

function

keyword

It can also check extend types

struct Ex: User

{:array, type} : array of type